+++ app/Http/Livewire/Admin/ListaProcesos.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire\Admin;
 
+use App\Models\Actividad;
 use App\Models\Proceso;
 use Livewire\Component;
 
@@ -25,6 +26,37 @@ class ListaProcesos extends Component
 
     public function eliminar($id)
     {
-        Proceso::find($id)->delete();
+        $dependientes = Proceso::query()
+            ->whereIn('id', function ($query) use ($id) {
+                $query->select('proceso_id')->from('actividades')
+                    ->where('proceso_id', $id);
+            })
+            ->orWhereIn('id', function ($query) use ($id) {
+                $query->select('proceso_id')->from('requisitos')
+                    ->where('proceso_id', $id);
+            })
+            ->orWhereIn('id', function ($query) use ($id) {
+                $query->select('proceso_id')->from('entradas')
+                    ->where('proceso_id', $id);
+            })
+            ->orWhereIn('id', function ($query) use ($id) {
+                $query->select('proceso_id')->from('encuesta_preguntas')
+                    ->where('proceso_id', $id);
+            })
+            ->orWhereIn('id', function ($query) use ($id) {
+                $query->select('proceso_id')->from('indicadores')
+                    ->where('proceso_id', $id);
+            })
+            ->orWhereIn('id', function ($query) use ($id) {
+                $query->select('proceso_id')->from('facultad_procesos')
+                    ->where('proceso_id', $id);
+            })
+            ->count();
+
+        if ($dependientes > 0) {
+            $this->emit('error', 'No es posible eliminar este proceso porque tiene tablas  dependientes.');
+        } else {
+            Proceso::find($id)->delete();
+        }
     }
 }
